add tsconfig.json and rename client->src
