+++ index.js
@@ -1,7 +1,7 @@
-const Config = require('./client/config')
-const { Scanner } = require('./client/scanner')
-const StatsDB = require('./client/statsdb')
-const Wallet = require('./client/wallet')
+const Config = require('src/config')
+const { Scanner } = require('src/scanner')
+const StatsDB = require('src/statsdb')
+const Wallet = require('src/wallet')
 const version = require('./package.json').version;
 
 module.exports = {
