+++ src/scanner.js
